+++ sdk/compute/azure-mgmt-compute/azure/mgmt/compute/v2018_04_01/models/_models_py3.py
@@ -5371,9 +5371,10 @@ class VirtualMachineScaleSetListWithLinkResult(msrest.serialization.Model):
 class VirtualMachineScaleSetManagedDiskParameters(msrest.serialization.Model):
     """Describes the parameters of a ScaleSet managed disk.
 
-    :param storage_account_type: Specifies the storage account type for the managed disk. Possible
-     values are: Standard_LRS, Premium_LRS, and StandardSSD_LRS. Possible values include:
-     "Standard_LRS", "Premium_LRS", "StandardSSD_LRS".
+    :param storage_account_type: Specifies the storage account type for the managed disk. Managed
+     OS disk storage account type can only be set when you create the scale set. Possible values
+     are: Standard_LRS, Premium_LRS, and StandardSSD_LRS. Possible values include: "Standard_LRS",
+     "Premium_LRS", "StandardSSD_LRS".
     :type storage_account_type: str or ~azure.mgmt.compute.v2018_04_01.models.StorageAccountTypes
     """