+++ sdk/containerregistry/azure-containerregistry/azure/containerregistry/_container_registry_client.py
@@ -4,11 +4,18 @@
 # Licensed under the MIT License.
 # ------------------------------------
 from typing import TYPE_CHECKING
-from azure.core.exceptions import ClientAuthenticationError, ResourceNotFoundError, ResourceExistsError
+from azure.core.exceptions import (
+    ClientAuthenticationError,
+    ResourceNotFoundError,
+    ResourceExistsError,
+    HttpResponseError,
+    map_error,
+)
 from azure.core.paging import ItemPaged
 
 from ._base_client import ContainerRegistryBaseClient
 from ._container_repository_client import ContainerRepositoryClient
+from ._generated.models import AcrErrors
 from ._helpers import _parse_next_link
 from ._models import DeletedRepositoryResult
 
@@ -29,7 +36,7 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         :returns: None
         :raises: None
         """
-        if not endpoint.startswith("https://"):
+        if not endpoint.startswith("https://") and not endpoint.startswith("http://"):
             endpoint = "https://" + endpoint
         self._endpoint = endpoint
         self._credential = credential
@@ -63,45 +70,65 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         n = kwargs.pop("page_size", None)
         last = kwargs.pop("last", None)
 
-        cls = kwargs.pop('cls', None)  # type: ClsType["_models.Repositories"]
-        error_map = {
-            401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
-        }
-        error_map.update(kwargs.pop('error_map', {}))
+        cls = kwargs.pop("cls", None)  # type: ClsType["_models.Repositories"]
+        error_map = {401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError}
+        error_map.update(kwargs.pop("error_map", {}))
         accept = "application/json"
 
         def prepare_request(next_link=None):
             # Construct headers
             header_parameters = {}  # type: Dict[str, Any]
-            header_parameters['Accept'] = self._client._serialize.header("accept", accept, 'str')
+            header_parameters["Accept"] = self._client._serialize.header(  # pylint: disable=protected-access
+                "accept", accept, "str"
+            )
 
             if not next_link:
                 # Construct URL
-                url = '/acr/v1/_catalog'
+                url = "/acr/v1/_catalog"
                 path_format_arguments = {
-                    'url': self._client._serialize.url("self._config.url", self._client._config.url, 'str', skip_quote=True),
+                    "url": self._client._serialize.url(  # pylint: disable=protected-access
+                        "self._config.url",
+                        self._client._config.url,
+                        "str",
+                        skip_quote=True,  # pylint: disable=protected-access
+                    ),
                 }
-                url = self._client._client.format_url(url, **path_format_arguments)
+                url = self._client._client.format_url(url, **path_format_arguments)  # pylint: disable=protected-access
                 # Construct parameters
                 query_parameters = {}  # type: Dict[str, Any]
                 if last is not None:
-                    query_parameters['last'] = self._client._serialize.query("last", last, 'str')
+                    query_parameters["last"] = self._client._serialize.query(  # pylint: disable=protected-access
+                        "last", last, "str"
+                    )
                 if n is not None:
-                    query_parameters['n'] = self._client._serialize.query("n", n, 'int')
+                    query_parameters["n"] = self._client._serialize.query(  # pylint: disable=protected-access
+                        "n", n, "int"
+                    )
 
-                request = self._client._client.get(url, query_parameters, header_parameters)
+                request = self._client._client.get(  # pylint: disable=protected-access
+                    url, query_parameters, header_parameters
+                )
             else:
                 url = next_link
                 query_parameters = {}  # type: Dict[str, Any]
                 path_format_arguments = {
-                    'url': self._client._serialize.url("self._config.url", self._client._config.url, 'str', skip_quote=True),
+                    "url": self._client._serialize.url(  # pylint: disable=protected-access
+                        "self._config.url",
+                        self._client._config.url,
+                        "str",
+                        skip_quote=True,  # pylint: disable=protected-access
+                    ),
                 }
-                url = self._client._client.format_url(url, **path_format_arguments)
-                request = self._client._client.get(url, query_parameters, header_parameters)
+                url = self._client._client.format_url(url, **path_format_arguments)  # pylint: disable=protected-access
+                request = self._client._client.get(  # pylint: disable=protected-access
+                    url, query_parameters, header_parameters
+                )
             return request
 
         def extract_data(pipeline_response):
-            deserialized = self._client._deserialize('Repositories', pipeline_response)
+            deserialized = self._client._deserialize(  # pylint: disable=protected-access
+                "Repositories", pipeline_response
+            )
             list_of_elem = deserialized.repositories
             if cls:
                 list_of_elem = cls(list_of_elem)
@@ -113,19 +140,21 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         def get_next(next_link=None):
             request = prepare_request(next_link)
 
-            pipeline_response = self._client._client._pipeline.run(request, stream=False, **kwargs)
+            pipeline_response = self._client._client._pipeline.run(  # pylint: disable=protected-access
+                request, stream=False, **kwargs
+            )
             response = pipeline_response.http_response
 
             if response.status_code not in [200]:
-                error = self._client._deserialize.failsafe_deserialize(_models.AcrErrors, response)
+                error = self._client._deserialize.failsafe_deserialize(  # pylint: disable=protected-access
+                    AcrErrors, response
+                )
                 map_error(status_code=response.status_code, response=response, error_map=error_map)
                 raise HttpResponseError(response=response, model=error)
 
             return pipeline_response
 
-        return ItemPaged(
-            get_next, extract_data
-        )
+        return ItemPaged(get_next, extract_data)
 
     def get_repository_client(self, repository, **kwargs):
         # type: (str, Dict[str, Any]) -> ContainerRepositoryClient