+++ sdk/containerregistry/azure-containerregistry/azure/containerregistry/aio/_async_container_registry_client.py
@@ -6,10 +6,17 @@
 from typing import Any, Dict, TYPE_CHECKING
 
 from azure.core.async_paging import AsyncItemPaged, AsyncList
-from azure.core.exceptions import ClientAuthenticationError, ResourceExistsError, ResourceNotFoundError
+from azure.core.exceptions import (
+    ClientAuthenticationError,
+    ResourceExistsError,
+    ResourceNotFoundError,
+    HttpResponseError,
+    map_error,
+)
 
 from ._async_base_client import ContainerRegistryBaseClient
 from ._async_container_repository_client import ContainerRepositoryClient
+from .._generated.models import AcrErrors
 from .._helpers import _parse_next_link
 from .._models import RepositoryProperties, DeletedRepositoryResult
 
@@ -18,10 +25,17 @@ if TYPE_CHECKING:
 
 
 class ContainerRegistryClient(ContainerRegistryBaseClient):
-    def __init__(
-        self, endpoint: str, credential: "AsyncTokenCredential", **kwargs
-    ):  # pylint: disable=client-method-missing-type-annotations
-        if not endpoint.startswith("https://"):
+    def __init__(self, endpoint: str, credential: "AsyncTokenCredential", **kwargs: Dict[str, Any]) -> None:
+        """Create a ContainerRegistryClient from an endpoint and a credential
+
+        :param endpoint: An ACR endpoint
+        :type endpoint: str
+        :param credential: The credential with which to authenticate
+        :type credential: AsyncTokenCredential
+        :returns: None
+        :raises: None
+        """
+        if not endpoint.startswith("https://") and not endpoint.startswith("http://"):
             endpoint = "https://" + endpoint
         self._endpoint = endpoint
         self._credential = credential
@@ -38,7 +52,7 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         result = await self._client.container_registry.delete_repository(repository, **kwargs)
         return DeletedRepositoryResult._from_generated(result)  # pylint: disable=protected-access
 
-    async def list_repositories(self, **kwargs) -> AsyncItemPaged[str]:
+    def list_repositories(self, **kwargs) -> AsyncItemPaged[str]:
         """List all repositories
 
         :keyword max: Maximum number of repositories to return
@@ -51,45 +65,65 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         n = kwargs.pop("page_size", None)
         last = kwargs.pop("last", None)
 
-        cls = kwargs.pop('cls', None)
-        error_map = {
-            401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
-        }
-        error_map.update(kwargs.pop('error_map', {}))
+        cls = kwargs.pop("cls", None)
+        error_map = {401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError}
+        error_map.update(kwargs.pop("error_map", {}))
         accept = "application/json"
 
         def prepare_request(next_link=None):
             # Construct headers
             header_parameters = {}  # type: Dict[str, Any]
-            header_parameters['Accept'] = self._client._serialize.header("accept", accept, 'str')
+            header_parameters["Accept"] = self._client._serialize.header(  # pylint: disable=protected-access
+                "accept", accept, "str"
+            )
 
             if not next_link:
                 # Construct URL
-                url = '/acr/v1/_catalog'
+                url = "/acr/v1/_catalog"
                 path_format_arguments = {
-                    'url': self._client._serialize.url("self._config.url", self._client._config.url, 'str', skip_quote=True),
+                    "url": self._client._serialize.url(  # pylint: disable=protected-access
+                        "self._config.url",
+                        self._client._config.url,
+                        "str",
+                        skip_quote=True,  # pylint: disable=protected-access
+                    ),
                 }
-                url = self._client._client.format_url(url, **path_format_arguments)
+                url = self._client._client.format_url(url, **path_format_arguments)  # pylint: disable=protected-access
                 # Construct parameters
                 query_parameters = {}  # type: Dict[str, Any]
                 if last is not None:
-                    query_parameters['last'] = self._client._serialize.query("last", last, 'str')
+                    query_parameters["last"] = self._client._serialize.query(  # pylint: disable=protected-access
+                        "last", last, "str"
+                    )
                 if n is not None:
-                    query_parameters['n'] = self._client._serialize.query("n", n, 'int')
+                    query_parameters["n"] = self._client._serialize.query(  # pylint: disable=protected-access
+                        "n", n, "int"
+                    )
 
-                request = self._client._client.get(url, query_parameters, header_parameters)
+                request = self._client._client.get(  # pylint: disable=protected-access
+                    url, query_parameters, header_parameters
+                )
             else:
                 url = next_link
                 query_parameters = {}  # type: Dict[str, Any]
                 path_format_arguments = {
-                    'url': self._client._serialize.url("self._config.url", self._client._config.url, 'str', skip_quote=True),
+                    "url": self._client._serialize.url(  # pylint: disable=protected-access
+                        "self._config.url",
+                        self._client._config.url,
+                        "str",
+                        skip_quote=True,  # pylint: disable=protected-access
+                    ),
                 }
-                url = self._client._client.format_url(url, **path_format_arguments)
-                request = self._client._client.get(url, query_parameters, header_parameters)
+                url = self._client._client.format_url(url, **path_format_arguments)  # pylint: disable=protected-access
+                request = self._client._client.get(  # pylint: disable=protected-access
+                    url, query_parameters, header_parameters
+                )
             return request
 
         async def extract_data(pipeline_response):
-            deserialized = self._client._deserialize('Repositories', pipeline_response)
+            deserialized = self._client._deserialize(
+                "Repositories", pipeline_response
+            )  # pylint: disable=protected-access
             list_of_elem = deserialized.repositories
             if cls:
                 list_of_elem = cls(list_of_elem)
@@ -101,19 +135,21 @@ class ContainerRegistryClient(ContainerRegistryBaseClient):
         async def get_next(next_link=None):
             request = prepare_request(next_link)
 
-            pipeline_response = await self._client._client._pipeline.run(request, stream=False, **kwargs)
+            pipeline_response = await self._client._client._pipeline.run(  # pylint: disable=protected-access
+                request, stream=False, **kwargs
+            )
             response = pipeline_response.http_response
 
             if response.status_code not in [200]:
-                error = self._client._deserialize.failsafe_deserialize(_models.AcrErrors, response)
+                error = self._client._deserialize.failsafe_deserialize(  # pylint: disable=protected-access
+                    AcrErrors, response
+                )
                 map_error(status_code=response.status_code, response=response, error_map=error_map)
                 raise HttpResponseError(response=response, model=error)
 
             return pipeline_response
 
-        return AsyncItemPaged(
-            get_next, extract_data
-        )
+        return AsyncItemPaged(get_next, extract_data)
 
     def get_repository_client(self, name: str, **kwargs) -> ContainerRepositoryClient: