+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_id_documents_async.py
@@ -26,7 +26,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
 
     @FormRecognizerPreparer()
     async def test_id_document_bad_endpoint(self, formrecognizer_test_endpoint, formrecognizer_test_api_key):
-        with open(self.id_document_jpg, "rb") as fd:
+        with open(self.id_document_license_jpg, "rb") as fd:
             myfile = fd.read()
         with self.assertRaises(ServiceRequestError):
             client = FormRecognizerClient("http://notreal.azure.com", AzureKeyCredential(formrecognizer_test_api_key))
@@ -43,7 +43,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
     @FormRecognizerPreparer()
     @GlobalClientPreparer()
     async def test_passing_enum_content_type(self, client):
-        with open(self.id_document_jpg, "rb") as fd:
+        with open(self.id_document_license_jpg, "rb") as fd:
             myfile = fd.read()
         async with client: 
             poller = await client.begin_recognize_id_documents(
@@ -108,7 +108,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
     @FormRecognizerPreparer()
     @GlobalClientPreparer()
     async def test_passing_bad_content_type_param_passed(self, client):
-        with open(self.id_document_jpg, "rb") as fd:
+        with open(self.id_document_license_jpg, "rb") as fd:
             myfile = fd.read()
         with self.assertRaises(ValueError):
             async with client:
@@ -147,7 +147,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
             responses.append(analyze_result)
             responses.append(extracted_id_document)
 
-        with open(self.id_document_jpg, "rb") as fd:
+        with open(self.id_document_license_jpg, "rb") as fd:
             myfile = fd.read()
 
         async with client:
@@ -177,8 +177,32 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
 
     @FormRecognizerPreparer()
     @GlobalClientPreparer()
+    async def test_id_document_jpg_passport(self, client):
+        with open(self.id_document_passport_jpg, "rb") as fd:
+            id_document = fd.read()
+
+        async with client:
+            poller = await client.begin_recognize_id_documents(id_document)
+
+            result = await poller.result()
+            self.assertEqual(len(result), 1)
+        
+            id_document = result[0]
+            # check dict values
+
+            passport = id_document.fields.get("MachineReadableZone").value
+            self.assertEqual(passport["LastName"].value, "MARTIN")
+            self.assertEqual(passport["FirstName"].value, "SARAH")
+            self.assertEqual(passport["DocumentNumber"].value, "ZE000509")
+            self.assertEqual(passport["DateOfBirth"].value, date(1985,1,1))
+            self.assertEqual(passport["DateOfExpiration"].value, date(2023,1,14))
+            self.assertEqual(passport["Sex"].value, "F")
+            self.assertEqual(passport["Country"].value, "CAN")
+
+    @FormRecognizerPreparer()
+    @GlobalClientPreparer()
     async def test_id_document_jpg(self, client):
-        with open(self.id_document_jpg, "rb") as fd:
+        with open(self.id_document_license_jpg, "rb") as fd:
             id_document = fd.read()
 
         async with client:
@@ -191,19 +215,19 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
 
         self.assertEqual(id_document.fields.get("LastName").value, "TALBOT")
         self.assertEqual(id_document.fields.get("FirstName").value, "LIAM R.")
-        # self.assertEqual(id_document.fields.get("DocumentNumber").value, "WDLABCD456DG") # service error when reading the license number returns 'LICWDLACD5DG'
+        # FIXME: service error when reading the license number returns 'LICWDLACD5DG'
+        # self.assertEqual(id_document.fields.get("DocumentNumber").value, "WDLABCD456DG")
         self.assertEqual(id_document.fields.get("DateOfBirth").value, date(1958,1,6))
         self.assertEqual(id_document.fields.get("DateOfExpiration").value, date(2020,8,12))
-        self.assertEqual(id_document.fields.get("Sex").value_data.text, "M")
+        self.assertEqual(id_document.fields.get("Sex").value, "M")
         self.assertEqual(id_document.fields.get("Address").value, "123 STREET ADDRESS YOUR CITY WA 99999-1234")
-        # FIXME: country is not returning a value
-        # self.assertEqual(id_document.fields.get("Country").value_data.text, "United States")
+        self.assertEqual(id_document.fields.get("Country").value, "USA")
         self.assertEqual(id_document.fields.get("Region").value, "Washington")
 
     @FormRecognizerPreparer()
     @GlobalClientPreparer()
     async def test_id_document_jpg_include_field_elements(self, client):
-        with open(self.id_document_jpg, "rb") as fd:
+        with open(self.id_document_license_jpg, "rb") as fd:
             id_document = fd.read()
         async with client:
             poller = await client.begin_recognize_id_documents(id_document, include_field_elements=True)
@@ -216,19 +240,18 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
 
         for field in id_document.fields.values():
             if field.name == "Country":
-                # FIXME: this is different than the other field values
-                # self.assertEqual(field.value_data.text, "United States")
+                self.assertEqual(field.value, "USA")
                 continue
             elif field.name == "Region":
                 self.assertEqual(field.value, "Washington")
-            else: 
+            else:
                 self.assertFieldElementsHasValues(field.value_data.field_elements, id_document.page_range.first_page_number)
 
     @FormRecognizerPreparer()
     @GlobalClientPreparer()
     @pytest.mark.live_test_only
     async def test_id_document_continuation_token(self, client):
-        with open(self.id_document_jpg, "rb") as fd:
+        with open(self.id_document_license_jpg, "rb") as fd:
             id_document = fd.read()
         async with client:
             initial_poller = await client.begin_recognize_id_documents(id_document)
@@ -241,7 +264,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
     @FormRecognizerPreparer()
     @GlobalClientPreparer(client_kwargs={"api_version": FormRecognizerApiVersion.V2_0})
     async def test_id_document_v2(self, client):
-        with open(self.id_document_jpg, "rb") as fd:
+        with open(self.id_document_license_jpg, "rb") as fd:
             id_document = fd.read()
         with pytest.raises(ValueError) as e:
             async with client:
@@ -251,7 +274,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
     @FormRecognizerPreparer()
     @GlobalClientPreparer()
     async def test_pages_kwarg_specified(self, client):
-        with open(self.id_document_jpg, "rb") as fd:
+        with open(self.id_document_license_jpg, "rb") as fd:
             id_document = fd.read()
         async with client:
             poller = await client.begin_recognize_id_documents(id_document, pages=["1"])