+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -1,13 +1,15 @@
 # Release History
 
-## 3.1.0b4 (Unreleased)
+## 3.1.0b4 (2021-04-06)
 
 **New features**
 
 - New methods `begin_recognize_id_documents` and `begin_recognize_id_documents_from_url` introduced to the SDK. Use these methods to recognize data from identity documents.
+- New field value types "gender" and "country" described in the `FieldValueType` enum.
 - Content-type `image/bmp` now supported by custom forms and training methods.
 - Added keyword argument `pages` for business cards, receipts, custom forms, and invoices 
 to specify which page to process of the document.
+- Added keyword argument `reading_order` to `begin_recognize_content` and `begin_recognize_content_from_url`.
 
 **Dependency Updates**